requirements.txt
# AbilityPath AI Screening Agent - Requirements
# Install with: pip install -r requirements.txt

# Core AI/LLM
langchain==0.1.0
langchain-openai==0.0.5
openai==1.12.0

# Web Framework  
flask==3.0.0
flask-cors==4.0.0

# Data handling
python-dotenv==1.0.0

# Optional but recommended
requests==2.31.0
